expose models as named exported variables
