+++ assets/src/components/form/select/model-selects/ticket.js
@@ -2,7 +2,7 @@
  * Internal imports
  */
 import ModelSelect from '../model-select';
-import * as ticketModel from '../../../../data/model/ticket';
+import { ticketModel } from '../../../../data/model';
 
 /**
  * External imports
