+++ assets/src/data/model/index.js
@@ -4,3 +4,4 @@ export * from './primary-keys';
 export * from './validators';
 export * from './model-names';
 export * from './base';
+export * from './models';
