+++ assets/src/data/model/models.js
@@ -1,0 +1,6 @@
+import * as dateTimeModel from './datetime';
+import * as eventModel from './event';
+import * as statusModel from './status';
+import * as ticketModel from './ticket';
+
+export { dateTimeModel, eventModel, statusModel, ticketModel };
